Fixed internal processing 2.

+++ lib/ui/warikan_calculator/warikan_calculator_body.dart
@@ -112,7 +112,7 @@ class WarikanCalculatorBody extends HookConsumerWidget {
 
   void _showResultDialog(
     BuildContext context,
-    double result,
+    int result,
   ) {
     showDialog(
       context: context,
